test: added unit test

+++ pyjira/services.py
@@ -1,10 +1,9 @@
 from typing import Sequence
 
-from sqlalchemy.exc import NoResultFound
 from sqlmodel import Session, select
 
 from pyjira.app import logger
-from pyjira.exceptions import RecordNotFoundError
+from pyjira import exceptions
 from pyjira.models import Task
 
 
@@ -14,6 +13,10 @@ class TaskService:
 
     def create_task(self, task_data: dict) -> Task:
         logger.info("Creating task.....")
+        if self.get_task(task_title=task_data["title"]):
+            raise exceptions.RecordAlreadyExistError(
+                message=f"Task already exist with title: {task_data['title']}"
+            )
         task = Task(
             title=task_data["title"],
             description=task_data["description"],
@@ -29,31 +32,25 @@ class TaskService:
 
     def get_task(self, task_title: str) -> Task:
         statement = select(Task).where(Task.title == task_title)
-        return self.session.exec(statement).one()
+        return self.session.exec(statement).one_or_none()
 
     def update_task(self, task_data: dict) -> Task:
-        try:
-            task = self.get_task(task_title=task_data["title"])
-            if task_data.get("description"):
-                task.description = task_data["description"]
-            if task_data.get("completed"):
-                task.completed = task_data["completed"]
-            self.session.commit()
-            self.session.refresh(task)
-            return task
-        except NoResultFound:
-            logger.error(f"Task not found with title: {task_data['title']}")
-            raise RecordNotFoundError(
+        if not (task := self.get_task(task_title=task_data["title"])):
+            raise exceptions.RecordNotFoundError(
                 message=f"Task not found with title: {task_data['title']}"
             )
+        if task_data.get("description"):
+            task.description = task_data["description"]
+        if task_data.get("completed"):
+            task.completed = task_data["completed"]
+        self.session.commit()
+        self.session.refresh(task)
+        return task
 
     def delete_task(self, task_title: str) -> None:
-        try:
-            task = self.get_task(task_title=task_title)
-            self.session.delete(task)
-            self.session.commit()
-        except NoResultFound:
-            logger.error(f"Task not found with title: {task_title}")
-            raise RecordNotFoundError(
+        if not (task := self.get_task(task_title=task_title)):
+            raise exceptions.RecordNotFoundError(
                 message=f"Task not found with title: {task_title}"
             )
+        self.session.delete(task)
+        self.session.commit()
